requirements.txt
# Function dependencies, for example:
# package>=version


google-cloud-monitoring>=2.19.2
google-cloud-pubsub>=2.19.7
google-api-core>=2.17.1
